refactor: transform calculations to base
