+++ tests/test_models.py
@@ -7,6 +7,8 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from eryx.models import *
 from eryx.pdb import AtomicModel
 from eryx.map_utils import generate_grid
+from eryx.base import compute_molecular_transform
+from eryx.base import compute_crystal_transform
 from base import setup_model
 
 class TestTransforms:
